Tests/LibWeb/Text/input/abspos-relayout-aspect-ratio.html

<!DOCTYPE html>
<script src="include.js"></script>
<style>
    body { margin: 0; }
    #root { position: relative; width: 400px; height: 300px; }
    #boundary {
        position: absolute;
        left: 20px;
        top: 10px;
        width: 150px;
        /* Height derives from the aspect ratio; replay re-solves it on every partial relayout. */
        aspect-ratio: 3 / 2;
        overflow: hidden;
    }
    #content { width: 50px; height: 30px; background: blue; }
</style>
<div id="root">
    <div id="boundary"><div id="content"></div></div>
</div>
<script>
    function rectToString(rect) {
        return `${rect.left},${rect.top},${rect.width},${rect.height}`;
    }

    asyncTest((done) => {
        document.body.offsetHeight;

        const boundary = document.getElementById("boundary");
        const partialBefore = internals.partialLayoutCount();
        const fullBefore = internals.fullLayoutCount();

        document.getElementById("content").style.width = "80px";
        document.body.offsetHeight;

        const rect = boundary.getBoundingClientRect();
        const content = document.getElementById("content").getBoundingClientRect();
        const partialDelta = internals.partialLayoutCount() - partialBefore;
        const fullDelta = internals.fullLayoutCount() - fullBefore;

        println(`boundary=${rectToString(rect)}`);
        println(`content=${rectToString(content)}`);
        println(`partial-layouts=${partialDelta} full-layouts=${fullDelta}`);
        println(rectToString(rect) === "20,10,150,100"
            && rectToString(content) === "20,10,80,30"
            && partialDelta === 1 && fullDelta === 0 ? "PASS" : "FAIL");
        done();
    });
</script>
